# packages/contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sophia application API",
    "version": "0.4.0-design",
    "description": "Sophia-owned binding contract. Not the Omnigent OpenAPI and not a claim that these routes are deployed."
  },
  "paths": {
...
    "/v1/runtime/hello": {
      "post": {
        "operationId": "runtimeHello",
        "summary": "A runtime instance opens (or takes over) its lease; the reply names the bindings it must own and where to resume reading commands.",
        "tags": [
          "runtime"
        ],
        "security": [
          {
            "runtimeCapability": []
          }
        ],
        "parameters": [
          {
            "name": "x-sophia-runtime-unit",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 160
            }
          },
          {
            "name": "x-sophia-bridge-instance",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          },
          {
            "name": "x-sophia-bridge-protocol",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeHello"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The lease this bridge instance now holds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeHelloReply"
                }
              }
            }
          },
          "401": {
            "description": "No or unknown runtime capability"
          },
          "403": {
            "description": "The capability is revoked or names another runtime unit"
          },
          "422": {
            "description": "Request fails the declared schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependent service unavailable"
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A"
      }
    },
    "/v1/runtime/commands": {
      "get": {
        "operationId": "runtimeCommands",
        "summary": "Commands admitted for this runtime after the cursor, in order; waits up to waitMs when there are none.",
        "tags": [
          "runtime"
        ],
        "security": [
          {
            "runtimeCapability": []
          }
        ],
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": true,
            "description": "The cursor: decimal digits of a safe integer",
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]{0,15})$"
            }
          },
          {
            "name": "waitMs",
            "in": "query",
            "required": true,
            "description": "How long to wait for a command, in milliseconds; at most 25000",
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]{0,4})$"
            }
          },
          {
            "name": "x-sophia-runtime-unit",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 160
            }
          },
          {
            "name": "x-sophia-bridge-instance",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          },
          {
            "name": "x-sophia-bridge-protocol",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A batch, possibly empty",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeCommandBatch"
                }
              }
            }
          },
          "401": {
            "description": "No or unknown runtime capability"
          },
          "403": {
            "description": "The capability is revoked or names another runtime unit"
          },
          "409": {
            "description": "Another bridge instance holds the lease; say hello again",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependent service unavailable"
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A"
      }
    },
    "/v1/runtime/receipts": {
      "post": {
        "operationId": "runtimeReceipts",
        "summary": "What the bridge observed for commands delivered to it. Idempotent per command and stage; a receipt for a command this runtime was never sent is refused.",
        "tags": [
          "runtime"
        ],
        "security": [
          {
            "runtimeCapability": []
          }
        ],
        "parameters": [
          {
            "name": "x-sophia-runtime-unit",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 160
            }
          },
          {
            "name": "x-sophia-bridge-instance",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          },
          {
            "name": "x-sophia-bridge-protocol",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeReceiptBatch"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Recorded (or already recorded)"
          },
          "401": {
            "description": "No or unknown runtime capability"
          },
          "403": {
            "description": "A receipt names a command, attempt or native session this runtime does not own"
          },
          "409": {
            "description": "Another bridge instance holds the lease",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request fails the declared schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependent service unavailable"
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A"
      }
    },
    "/v1/runtime/observations": {
      "post": {
        "operationId": "runtimeObservations",
        "summary": "Durable native session events of attempts bound to this runtime, once per (runtime unit, native session, native seq). Observations are evidence: they never act as a member or accept work.",
        "tags": [
          "runtime"
        ],
        "security": [
          {
            "runtimeCapability": []
          }
        ],
        "parameters": [
          {
            "name": "x-sophia-runtime-unit",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 160
            }
          },
          {
            "name": "x-sophia-bridge-instance",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          },
          {
            "name": "x-sophia-bridge-protocol",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeObservationBatch"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Recorded (or already recorded)"
          },
          "401": {
            "description": "No or unknown runtime capability"
          },
          "403": {
            "description": "An observation names an attempt or native session this runtime does not own"
          },
          "409": {
            "description": "Another bridge instance holds the lease",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request fails the declared schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependent service unavailable"
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "supabaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "runtimeCapability": {
        "type": "http",
        "scheme": "bearer",
        "description": "A runtime capability: an opaque secret held by the execution host and stored by the service only as SHA-256. It is bound to one project, executor resource and runtime unit, and it is not a member identity."
      },
      "mediaBridgeCapability": {
        "type": "http",
        "scheme": "bearer",
        "description": "The media bridge's capability: an opaque secret on the bridge host; the API holds only its SHA-256. It names no member and grants no member authority: tool calls act for the speaker the exchange's input epoch binds, under that speaker's own role."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "retry": {
            "type": "string",
            "enum": [
              "safe_read",
              "same_admission_key",
              "reconcile_first",
              "reauthorize",
              "never"
            ]
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "retry"
        ]
      },
...
      "RuntimeReceipt": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "commandId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$"
          },
          "attemptId": {
            "type": "string",
            "pattern": "^([A-Za-z0-9][A-Za-z0-9_-]{0,63})?$"
          },
          "stage": {
            "type": "string",
            "enum": [
              "delivered",
              "incorporation_observed",
              "checked",
              "rejected",
              "failed",
              "outcome_unknown"
            ]
          },
          "nativeSessionId": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,159}$"
              },
              {
                "type": "null"
              }
            ]
          },
          "nativeSequence": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "evidenceRefs": {
            "type": "array",
            "maxItems": 16,
            "items": {
              "type": "string",
              "maxLength": 256
            }
          },
          "observedAt": {
            "type": "string",
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9:.]+Z$"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16000
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "commandId",
          "attemptId",
          "stage",
          "nativeSessionId",
          "nativeSequence",
          "evidenceRefs",
          "observedAt",
          "reason"
        ]
      },
      "RuntimeObservation": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "runtimeUnitId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,159}$"
          },
          "attemptId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$"
          },
          "nativeSessionId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,159}$"
          },
          "nativeSeq": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "type": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_/.-]{0,63}$"
          },
          "durable": {
            "const": true
          },
          "data": {}
        },
        "required": [
          "runtimeUnitId",
          "attemptId",
          "nativeSessionId",
          "nativeSeq",
          "type",
          "durable",
          "data"
        ]
      },
      "RuntimeHello": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "bundle": {
            "type": "string",
            "maxLength": 200
          },
          "protocolVersion": {
            "const": 1
          },
          "dshVersion": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "bundle",
          "protocolVersion",
          "dshVersion"
        ]
      },
      "RuntimeServiceBinding": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "attemptId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$"
          },
          "nativeSessionId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,159}$"
          },
          "authorityEpoch": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "state": {
            "type": "string",
            "enum": [
              "active",
              "held",
              "stopped"
            ]
          }
        },
        "required": [
          "attemptId",
          "nativeSessionId",
          "authorityEpoch",
          "state"
        ]
      },
      "RuntimeHelloReply": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "projectId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$"
          },
          "leaseId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$"
          },
          "authorityEpoch": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "bindings": {
            "type": "array",
            "maxItems": 500,
            "items": {
              "$ref": "#/components/schemas/RuntimeServiceBinding"
            }
          },
          "cursor": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "projectId",
          "leaseId",
          "authorityEpoch",
          "bindings",
          "cursor"
        ]
      },
      "RuntimeCommandBatch": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "commands": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "seq": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 9007199254740991
                },
                "command": {}
              },
              "required": [
                "seq",
                "command"
              ]
            }
          },
          "cursor": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "commands",
          "cursor"
        ]
      },
      "RuntimeReceiptBatch": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "receipts": {
            "type": "array",
            "minItems": 1,
            "maxItems": 500,
            "items": {
              "$ref": "#/components/schemas/RuntimeReceipt"
            }
          }
        },
        "required": [
          "receipts"
        ]
      },
      "RuntimeObservationBatch": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "observations": {
            "type": "array",
            "minItems": 1,
            "maxItems": 500,
            "items": {
              "$ref": "#/components/schemas/RuntimeObservation"
            }
          }
        },
        "required": [
          "observations"
        ]
      },
...
    }
  }
}